Succesffully Getting Indices And Weights Again
Forgot to make the weight a float, and the index was commented out.

+++ FBX_DLL/FBX_Helper_Structures.h
@@ -10,7 +10,7 @@ enum Vertex_Part_Data { X = 0, Y = 1, Z = 2 };
 struct BlendingIndexWeightPair
 {
 	unsigned int pBlendingIndex;
-	unsigned int pBlendingWeight;
+	float pBlendingWeight;
 };
 
 struct CtrlPoint
